Added a metamap to connect the worldmaps. Rearranged a lot of stuff.. The level name that's stored in the StatManager is now the actual levelname from the .pingus file (unless blank), so the old savegames.xml file won't be effective anymore.
Also added some things to the variables.xml file.

More general cleanup, too.  Big commit...

+++ src/worldmap/pingus.cxx
@@ -121,7 +121,7 @@ float
 Pingus::get_direction() const
 {
   return (atan2(last_pos.x - pos.x,
-                -(last_pos.y - pos.y)) / Math::pi * 180.0f) + 180.0f;
+                -(last_pos.y - pos.y)) / (float)Math::pi * 180.0f) + 180.0f;
 }
 
 bool
