+++ src/worldobjs/fake_exit.cxx
@@ -34,7 +34,7 @@ FakeExit::FakeExit(const FileReader& reader)
     smashing(false)
 {
   reader.read_vector("position", pos);
-  pos -= Vector(surface.get_width ()/2, surface.get_height ());
+  pos -= Vector((float)surface.get_width ()/2, (float)surface.get_height ());
 }
 
 float
